[update] 커뮤니티 알림 모달 적용 완 / 수정 삭제 제어 완

+++ src/main/webapp/WEB-INF/views/board/read.jsp
@@ -46,6 +46,28 @@
 <script type="text/javascript" src="${pageContext.request.contextPath}/resources/js/comment.js"></script>
 <script type="text/javascript">
 $(document).ready(function(){
+	// 알림 모달창 호출 함수 --------------
+	function success(result) {
+	    Swal.fire(
+	        result,
+	        '',
+	        'success' /*디자인 타입*/
+	    )
+	}//success 버튼
+	
+	//warning 버튼
+	function warning(result) {
+        Swal.fire(
+            result,
+            '',
+            'warning' /*디자인 타입*/
+        )
+    }//warning 버튼
+	// 알림 모달창 호출 함수 끝 --------------
+	
+	
+	
+	
 	// 계속 쓸 놈들 ---------------------
 	var loginID = '<c:out value="${loginID}"/>';
 	var bnoValue = '<c:out value="${vo.bno}"/>';
@@ -72,7 +94,7 @@ $(document).ready(function(){
 			// 반복문 돌면서 댓글 list 채우기
 			for (var i = 0, len = list.length||0; i < len; i++) {
 				str += "<li id='cmtLI' data-cno='"+list[i].cno+"' class='mb-5'>";
-				str += "<div id='cmt-body'><div id='cmt-header'><strong>" + list[i].nick +" /cno: " + list[i].cno + "</strong>&nbsp;&nbsp;&nbsp;";
+				str += "<div id='cmt-body'><div id='cmt-header'><strong>" + list[i].nick + "</strong>&nbsp;&nbsp;&nbsp;";
 					// 수정 날짜 null이면 regdate로 표시, 수정 날짜 null 아니면 updatedate로 표시
 					if (list[i].c_updatedate == null){
 						str += "<span><small> "+cmtService.displayTime(list[i].c_regdate)+"</small></span><span>";
@@ -80,11 +102,11 @@ $(document).ready(function(){
 						str += "<span><small> "+cmtService.displayTime(list[i].c_updatedate)+"</small></span><span>";
 					}
 					
-					// id가 admin이거나 본인일 때만 -> 답글, 수정, 삭제 버턴 나오게 제어
+						str += "<button id='cmt_btn_re' style='border: none; margin: 2px; background-color: #ecdffd; border-radius: 5px; float: right;'> 답글</button>";
+					// id가 admin이거나 본인일 때만 -> 수정, 삭제 버턴 나오게 제어
 					if (list[i].id == loginID || loginID == 'admin') {
 						str += "<button id='cmt_btn_del' style='border: none; margin: 2px; background-color: #ecdffd; border-radius: 5px; float: right;'>삭제</button>";
 						str += "<button id='cmt_btn_mod' style='border: none; margin: 2px; background-color: #ecdffd; border-radius: 5px; float: right;'>수정</button>";
-						str += "<button id='cmt_btn_re' style='border: none; margin: 2px; background-color: #ecdffd; border-radius: 5px; float: right;'> 답글</button>";
 						str += "<input type='hidden' value='"+list[i].cno+"' id='cnoValue'></span></div>";
 					}
 				str += "<div id='c_contentDIV'>" + list[i].c_content+"</div>";
@@ -105,7 +127,7 @@ $(document).ready(function(){
 	cmtRegisterBtn.on("click", function(e){
 		
 		if($('#c_content').val() == null || $('#c_content').val() == ''){
-			alert("댓글 내용을 작성해주세요");
+			warning("댓글 내용을 작성해주세요");
 			$('#c_content').focus();
 			return false;
 		}
@@ -127,7 +149,7 @@ $(document).ready(function(){
 					console.log("addResult: " + addResult);
 					
 					if(addResult === "success") {
-						alert("댓글이 등록되었습니다 👍 ");
+						success("댓글이 등록되었습니다 👍 ");
 					}
 					
 					showCmtList(1);
@@ -143,33 +165,69 @@ $(document).ready(function(){
 	
 	// 댓글 삭제 -------------------------------
 // 	cmtDelBtn.on("click", function(e){ // 버턴이 먹지를 않노 ㄱ-
-	$(document).on("click", "#cmt_btn_del", function(){
+	$(document).on("click", "#cmt_btn_del", function(){ // ㄴ이렇게 하니까 됨
 // 		alert("삭제 버턴 클릭됨");
 		// cno는?  삭제 버턴(this) -> 다음 다음 다음 요소의 value값.. 이게 최선?ㅠ
 		var cnoValue = $(this).next().next().next().val();
 		console.log("삭제할 cnoValue: " + cnoValue);
-		alert("삭제할 cnoValue: " + cnoValue);
 		
 		// 삭제 버튼 클릭했을 때~ 
-		// 댓글 삭제 함수 3. deleteCmt(cno, callback, error) 호출
-		cmtService.deleteCmt(
-			// cno
-			cnoValue,
-			// callback
-			function(deleteResult){
-				console.log("deleteResult: " + deleteResult);
-		
-				if(deleteResult === "success") {
-					alert("댓글이 삭제되었습니다 👍 ");
-				}
-				
-				showCmtList(1);
-			}, 
-			// error
-			function(error){
-				alert("삭제 실패...... ");
-			}
-		);// 3.deleteCmt()
+		// 삭제하시겠습니까? 알림창 띄우고 -> 예 클릭 시,, 삭제 함수 호출까지
+		confirmCmtDel();
+			
+			// 이 함수 선언을 밖에 하니까 cnoValue를 못 읽어와서 에러 뜨네;
+			//confirm(예, 아니오 확인) 버튼
+			function confirmCmtDel() {
+				Swal.fire({
+		           title: '선택하신 댓글을 <br> 삭제하시겠습니까?',
+		           text: '',
+		           icon: 'question',
+		           showCancelButton: true, //cancel 버튼 보이기
+		           confirmButtonColor: '#3085d6', //confirm 버튼 색깔
+		           cancelButtonColor: '#d33', //cancel 버튼 색깔
+		           confirmButtonText: '예',
+		           cancelButtonText: '아니요',
+		           reverseButtons: false //버튼 순서 거꾸로 
+		       }).then((result) => {
+		           if (result.isConfirmed) {
+		               // alert('cnoValue 들어오나? ' + cnoValue);
+		               // 예 눌렀을 시 -> 댓글 삭제 함수 호출
+		               			// 댓글 삭제 함수 3. deleteCmt(cno, callback, error) 호출
+								cmtService.deleteCmt(
+									// cno
+									cnoValue,
+									// callback
+									function(deleteResult){
+										console.log("deleteResult: " + deleteResult);
+								
+										if(deleteResult === "success") {
+											console.log("댓글이 삭제되었습니다");
+										}
+										
+										showCmtList(1);
+									}, 
+									// error
+									function(error){
+										warning("삭제 실패...... ");
+										//location.reload();
+									}
+								);// 3.deleteCmt()
+		           	
+		           	Swal.fire(
+		                   '댓글 삭제 완료 👍',
+		                   '',
+		                   'success'
+		               )
+		           } else if (result.isDismissed) {
+		               Swal.fire(
+		                   '댓글 삭제 취소 😮',
+		                   '',
+		                   'info'
+		               )
+		           } // if else
+		       })
+		   }//confirm(예, 아니오 확인) 버튼
+		   
 	}); // cmtDelBtn on click
 	// 댓글 삭제 끝 -------------------------------
 	
@@ -193,7 +251,7 @@ $(document).ready(function(){
 	// 버튼 클릭 이벤트
 	$(document).on("click", "#cmt_btn_mod", function(){
 // 		alert("수정 버턴 클릭");
-		var cnoValue = $(this).next().next().val();
+		var cnoValue = $(this).next().val();
 		var c_content = $("[data-cno="+cnoValue+"]").find('#c_contentDIV').text();
 
 // 		alert("cnoValue: " + cnoValue + " / c_content: " + c_content);
@@ -221,7 +279,7 @@ $(document).ready(function(){
 					
 					// callback
 					function(rData){
-						alert("댓글 수정 완");
+						  success("댓글 수정 완료💖");
 						
 						showCmtList(1);
 					}
@@ -232,7 +290,6 @@ $(document).ready(function(){
 	});// on click
 	// 댓글 수정 끝 -------------------------------
 	
-	
 }); // jquery ready
 
 </script>
@@ -240,15 +297,15 @@ $(document).ready(function(){
 
 
 
-
-
 		<!-- 수정, 삭제 시 필요한 글 번호(bno) 저장하는 폼태그 =====================-->
 		<form role="bno_form" method="post">
 			<input type="hidden" name="bno" value="${vo.bno }">
+			<input type="hidden" name="id" value="${vo.id}" readonly>
 		</form>
 		<!-- 수정, 삭제 시 필요한 글 번호(bno) 저장하는 폼태그 껏 =====================-->
-			<input type="hidden" name="id" value="${vo.id}" readonly>
 
+
+<!-- ============================= 본문 시작 ============================= -->
 <div class="container mt-5 ">
 	<div class="row" style="justify-content: center; align-items: center;">
 		<div class="col-lg-8">
@@ -257,7 +314,7 @@ $(document).ready(function(){
 				<!-- Post header-->
 				<header class="mb-4">
 					<!-- Post title-->
-					<h2 class="fw-bolder mb-1"> <span style="font-size: medium;">[ ${vo.b_ctgr} ]</span> ${vo.b_title }</h2>
+					<h3 class="fw-bolder mb-1"> <span style="font-size: medium;">[ ${vo.b_ctgr} ]</span> ${vo.b_title }</h3>
 					<!-- Post meta content-->
 					<div class="text-muted fst-italic mb-2">
 						<b>작성일시: </b> 						
@@ -266,21 +323,24 @@ $(document).ready(function(){
 							<c:otherwise>${vo.b_updatedate }</c:otherwise>
 						</c:choose> &nbsp;&nbsp;&nbsp; 
 						<b>작성자: </b> ${vo.nick }</div>
-						
-						
 				</header>
 				<!-- Preview image figure-->
 				<figure class="mb-4">
 <!-- 					<img class="img-fluid rounded" src="https://dummyimage.com/900x400/ced4da/6c757d.jpg" alt="..." /> -->
 				</figure>
 				<!-- Post content-->
-				<section class="mb-5">
+				<section class="card-body border-left-info shadow h-100 py-2 mb-5" style="padding-left: 5%;">
 					<p class="fs-5 mb-4">${vo.b_content }</p>
 				</section>
-				<button class="btn_mod"
-					style="border: none; margin: 2px; background-color: #5107B0; color: white; border-radius: 5px;">수정</button>
-				<button class="btn_del"
-					style="border: none; margin: 2px; background-color: #5107B0; color: white; border-radius: 5px;">삭제</button>
+				
+				<c:if test="${vo.id eq sessionScope.loginID}">
+					<button class="btn_mod"
+						style="border: none; margin: 2px; background-color: #5107B0; color: white; border-radius: 5px;">수정</button>
+				</c:if>
+				<c:if test="${vo.id eq sessionScope.loginID || sessionScope.loginID eq 'admin' }">
+					<button class="btn_del"
+						style="border: none; margin: 2px; background-color: #5107B0; color: white; border-radius: 5px;">삭제</button>
+				</c:if>
 				<button class="btn_list"
 					style="border: none; margin: 2px; background-color: #5107B0; color: white; border-radius: 5px;">목록</button>
 			</article>
@@ -304,7 +364,7 @@ $(document).ready(function(){
 						<!-- 댓글입력창 !-->
 						<form class="mb-4">
 							<textarea class="form-control" rows="3" name="content"
-								id="c_content" placeholder="댓글을 입력해주세요 !"></textarea>
+								id="c_content" placeholder="댓글을 입력해주세요🥰 "></textarea>
 							<button type="button" class="btn" id="add_cmt_btn"
 								style="background-color: #5107B0; float: right; width: 80px; margin: 10px; font-size: 16px;">
 								<span class="btn-inner-text" style="color: white;">등 록</span>
@@ -348,55 +408,7 @@ $(document).ready(function(){
     </div>
   </div>
 
-						<!-- ----------------------- 댓글 리스트 구간 --------------------------------- -->
-<!-- 						<div style="border: 1px solid black"> -->
-<!-- 							<div> -->
-<!-- 								<h3>댓글</h3> -->
-<!-- 							</div> -->
-<!-- 							<ul id="cmtUL"> -->
-<!-- 								<li data-cno=""> -->
-<!-- 									<div id="cmt-body"> -->
-<!-- 										<div id="cmt-header"> -->
-<!-- 											<strong> id,, 말고 nick </strong> <small> c_regdate </small> <input -->
-<!-- 												type="button" value="답글" class="btn" id="cmt_btn_re"> -->
-<!-- 											<input type="button" value="수정" class="btn" id="cmt_btn_mod"> -->
-<!-- 											<input type="button" value="삭제" class="btn" id="cmt_btn_del"> -->
-<!-- 											<input type="hidden" value="" id="cnoValue"> -->
-<!-- 										</div> -->
-<!-- 										<p>c_content</p> -->
-<!-- 									</div> -->
-<!-- 								</li> -->
-<!-- 							</ul> -->
-<!-- 						</div> -->
-						<!-- ----------------------- 댓글 리스트 구간 끝^^ --------------------------------- -->
-
-
-
-
-						<!-- ----------------------- 댓글 작성 구간^^ --------------------------------- -->
-						<!-- 						<div style="border: 1px solid black;"> -->
-						<!-- 							<h3>댓글을 남겨주세요 👇👇</h3> -->
-						<!-- 							<div class="form-group"> -->
-						<!-- 								<label for="message">내용</label> -->
-						<!-- 								<textarea name="content" id="c_content" cols="30" rows="5" -->
-						<!-- 									class=""></textarea> -->
-						<!-- 							</div> -->
-						<!-- 							<div class="btn btn-primary"> -->
-						<!-- 								<input type="button" value="댓글 달기😘" class="btn btn-primary" -->
-						<!-- 									id="add_cmt_btn"> -->
-						<!-- 							</div> -->
-						<!-- 						</div> -->
-
-						<!-- ----------------------- 댓글 작성 구간 끝^^ --------------------------------- -->
-
-
-
-
-
-
-
-						<script type="text/javascript">
-
+<script type="text/javascript">
 	// jQuery 구간 시작~ =================================================
 	$(document).ready(function() {
 		// 	  alert('jQuery 실행🎊🎊');
@@ -406,7 +418,7 @@ $(document).ready(function(){
 		var fr = $('form[role="bno_form"]'); // role이 form인 폼태그에 접근해서 그걸 fr 변수에 담기
 
 		$(".btn_mod").click(function() {
-			// 		  alert('정상 동작 중');
+			// alert('정상 동작 중');
 			// bno폼태그 속성 바꿀 거!!
 			fr.attr("action", "/board/update");
 			fr.attr("method", "get"); // get방식으로 바꿔서 전달 
@@ -421,13 +433,46 @@ $(document).ready(function(){
 
 		
 		$(".btn_del").click(function() {
-			// 삭제 버턴 클릭했을 때~~
-			alert('삭제 버턴 클릭됨');
-			fr.attr("action", "/board/delete");
-			fr.submit(); // 클릭하면? 속성 바꾸고 -> submit 되게~~
+			// 삭제 버턴 클릭했을 때~~ 알림창으로 한번 더 물어보고 예 누르면 삭제하는 주소로 보내기
+			// alert('삭제 버턴 클릭됨');
+			confirmBoardDel();
 		});// 삭제 버튼 click
-
+		
+			//confirm(예, 아니오 확인) 버튼
+			function confirmBoardDel() {
+		       Swal.fire({
+		           title: '선택하신 글을 <br> 삭제하시겠습니까?',
+		           text: '',
+		           icon: 'question',
+		           showCancelButton: true, //cancel 버튼 보이기
+		           confirmButtonColor: '#3085d6', //confirm 버튼 색깔
+		           cancelButtonColor: '#d33', //cancel 버튼 색깔
+		           confirmButtonText: '예',
+		           cancelButtonText: '아니요',
+		           reverseButtons: false //버튼 순서 거꾸로 
+		       }).then((result) => {
+		           if (result.isConfirmed) {
+		               // 예 눌렀을 시 -> 삭제 주소로 보내버리기
+		               	fr.attr("action", "/board/delete");
+						fr.submit(); // 클릭하면? 속성 바꾸고 -> submit 되게~~
+		           	
+		           	Swal.fire(
+		                   '삭제 완료 👍',
+		                   '',
+		                   'success'
+		               )
+		               
+		           } else if (result.isDismissed) {
+		               Swal.fire(
+		                   '삭제 취소 😮',
+		                   '',
+		                   'info'
+		               )
+		           }
+		       })
+		   }//confirm(예, 아니오 확인) 버튼
+		
 	});// jQuery ready
 </script>
 
-						<%@ include file="../include/footer.jsp"%>
+<%@ include file="../include/footer.jsp"%>
